Implement more code review suggestions
