+++ package.json
@@ -54,6 +54,7 @@
     "resolve-pathname": "^3.0.0",
     "styled-components": "^4.1.3",
     "web3": "^1.2.2",
+    "web3-eth-abi": "^1.2.4",
     "web3-utils": "^1.2.2"
   },
   "devDependencies": {
