+++ src/apps/Console/CommandHandlers.js
@@ -1,0 +1,123 @@
+import { hash as namehash } from 'eth-ens-namehash'
+import abi from 'web3-eth-abi'
+import {
+  parseInitParams,
+  parsePermissions,
+  parseMethodCall,
+} from './console-utils'
+import { encodeFunctionCallFromSignature } from './web3-encoding-utils'
+
+// Maximum time (in milliseconds) to wait for a response
+// from the aragon.js apm repo content fetcher
+const REPO_FETCH_TIMEOUT = 3000
+const APP_POSTFIX = '.aragonpm.eth'
+
+export async function getInstallTransactionPath(wrapper, apps, params) {
+  // Get & properly parse arguments
+  const [appName, initArgs, permArgs] = params
+  const parsedInitArgs = parseInitParams(initArgs)
+  const parsedPermArgs = parsePermissions(permArgs)
+  // Resolve namehash and ens domain to fetch the app's repo content
+  const appId = namehash(`${appName}${APP_POSTFIX}`)
+  const ensDomain = await wrapper.ens.resolve(appId)
+  const {
+    abi: appAbi,
+    contractAddress,
+    roles,
+  } = await wrapper.apm.fetchLatestRepoContent(ensDomain, {
+    fetchTimeout: REPO_FETCH_TIMEOUT,
+  })
+  // get the initialize function to properly install the app
+  const { name, type, inputs } = appAbi.find(
+    ({ name }) => name === 'initialize'
+  )
+
+  const functionObject = {
+    name: name,
+    type: type,
+    inputs: inputs,
+  }
+  const encodedInitializeFunc = abi.encodeFunctionCall(
+    functionObject,
+    parsedInitArgs
+  )
+
+  const kernelProxyAddress = apps.find(
+    app => app.name.toLowerCase() === 'kernel'
+  ).proxyAddress
+
+  const path = await wrapper.getTransactionPath(
+    kernelProxyAddress,
+    'newAppInstance(bytes32,address,bytes,bool)',
+    [appId, contractAddress, encodedInitializeFunc, false]
+  )
+
+  // Get the second to last item in the path, as it is the account that will execute kernel.newAppInstance
+  const scriptExecutor = path[path.length - 2].to
+  const counterfactualAppAddr = await wrapper.kernelProxy.call(
+    'newAppInstance',
+    appId,
+    contractAddress,
+    encodedInitializeFunc,
+    false,
+    { from: scriptExecutor }
+  )
+
+  const installAppIntent = [
+    [
+      kernelProxyAddress,
+      'newAppInstance(bytes32,address,bytes,bool)',
+      [appId, contractAddress, encodedInitializeFunc, false],
+    ],
+  ]
+
+  const aclProxyAddress = apps.find(app => app.name.toLowerCase() === 'acl')
+    .proxyAddress
+
+  const permissionIntents = parsedPermArgs.map(([role, from, to]) => {
+    const roleBytes = roles.find(availableRole => availableRole.id === role)
+      .bytes
+
+    return [
+      aclProxyAddress,
+      'createPermission',
+      [to, counterfactualAppAddr, roleBytes, from],
+    ]
+  })
+
+  const intentBasket = [...installAppIntent, ...permissionIntents]
+  const {
+    path: pathForBasket,
+  } = await wrapper.getTransactionPathForIntentBasket(intentBasket)
+
+  return pathForBasket
+}
+
+export async function getExecTransactionPath(wrapper, params) {
+  const [proxyAddress, methodWithArgs] = params
+  const [methodSignature, args] = parseMethodCall(methodWithArgs)
+  const path = await wrapper.getTransactionPath(
+    proxyAddress,
+    methodSignature,
+    args
+  )
+  return path
+}
+
+export async function getActTransactionPath(wrapper, params) {
+  const [selectedAgentInstance, targetAddress, methodWithArgs] = params
+  const [methodName, methodParams, methodArgs] = parseMethodCall(methodWithArgs)
+  const methodSignature = `${methodName}(${methodParams.join(',')})`
+
+  const encodedFunctionCall = encodeFunctionCallFromSignature(
+    methodSignature,
+    methodArgs
+  )
+
+  const path = await wrapper.getTransactionPath(
+    selectedAgentInstance,
+    'execute(address,uint256,bytes)',
+    [targetAddress, 0, encodedFunctionCall]
+  )
+  return path
+}
